# Advanced Database Connection Management and Batch Operations