--- a/src/main/java/org/hasting/util/DatabaseConnectionManager.java
+++ b/src/main/java/org/hasting/util/DatabaseConnectionManager.java
@@ -1,79 +1,203 @@
 package org.hasting.util;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
+import org.hasting.util.logging.MP3OrgLoggingManager;
+import org.hasting.util.logging.Logger;
+
+import java.sql.*;
 
 /**
- * Manages database connection testing and lock detection for the MP3Org application.
+ * Improved database connection manager that ensures connections are properly closed after use.
+ * This class provides methods that automatically manage connection lifecycle.
  * 
- * <p>This class provides utilities to test database availability without interfering
- * with the main application's database operations. It specializes in detecting Derby
- * database lock conditions that occur when multiple application instances attempt
- * to access the same database simultaneously.
- * 
- * <p><strong>Key Capabilities:</strong>
+ * <p>Key features:
  * <ul>
- *   <li>Quick database availability testing with minimal resource usage</li>
- *   <li>Derby-specific lock exception recognition and handling</li>
- *   <li>Non-invasive connection testing that doesn't affect primary database usage</li>
- *   <li>Clear error reporting for different database lock scenarios</li>
+ *   <li>Automatic connection closing after each operation</li>
+ *   <li>Explicit commit for all write operations</li>
+ *   <li>Proper rollback on errors</li>
+ *   <li>Connection validation before use</li>
  * </ul>
- * 
- * <p>This class follows the principle of self-documenting code where method names
- * clearly communicate their purpose and the implementation teaches the database
- * lock detection pattern to future developers.
- * 
- * @author MP3Org Development Team
- * @version 1.0
- * @since 1.0
- * @see DatabaseProfileManager
- * @see DatabaseManager
  */
 public class DatabaseConnectionManager {
-    
+    private static final Logger logger = MP3OrgLoggingManager.getLogger(DatabaseConnectionManager.class);
     private static final int CONNECTION_TIMEOUT_SECONDS = 5;
     
     /**
-     * Tests if a Derby database at the given path is currently locked by another process.
+     * Creates a new database connection that must be closed after use.
      * 
-     * <p>This method performs a quick connection test without creating tables or
-     * performing any database operations that might interfere with existing connections.
-     * It specifically looks for Derby database lock exceptions that indicate another
-     * process is currently using the database.
+     * @param config the database configuration
+     * @return a new database connection
+     * @throws SQLException if connection creation fails
+     */
+    public static Connection createConnection(DatabaseConfig config) throws SQLException {
+        Connection conn = DriverManager.getConnection(
+            config.getJdbcUrl(),
+            config.getUsername(), 
+            config.getPassword()
+        );
+        
+        // Set auto-commit to false for explicit transaction control
+        conn.setAutoCommit(false);
+        
+        return conn;
+    }
+    
+    /**
+     * Validates if a connection is still usable.
      * 
-     * <p><strong>Lock Detection Strategy:</strong>
-     * <ol>
-     *   <li>Attempt a brief connection to the database</li>
-     *   <li>Immediately close the connection if successful</li>
-     *   <li>Analyze any SQLException for Derby-specific lock indicators</li>
-     *   <li>Return true if lock detected, false if database is available</li>
-     * </ol>
+     * @param conn the connection to validate
+     * @return true if the connection is valid, false otherwise
+     */
+    public static boolean isConnectionValid(Connection conn) {
+        try {
+            return conn != null && !conn.isClosed() && conn.isValid(CONNECTION_TIMEOUT_SECONDS);
+        } catch (SQLException e) {
+            return false;
+        }
+    }
+    
+    /**
+     * Executes a database operation with automatic connection management.
+     * This method ensures the connection is properly closed after use.
      * 
-     * @param databasePath the file system path to the Derby database directory
-     * @return true if the database is locked by another process, false if available
-     * @throws IllegalArgumentException if databasePath is null or empty
+     * @param config the database configuration
+     * @param operation the operation to execute
+     * @param <T> the return type
+     * @return the result of the operation
+     * @throws SQLException if the operation fails
+     */
+    public static <T> T executeWithNewConnection(DatabaseConfig config, ConnectionOperation<T> operation) 
+            throws SQLException {
+        Connection conn = null;
+        boolean success = false;
+        
+        try {
+            conn = createConnection(config);
+            T result = operation.execute(conn);
+            conn.commit(); // Explicitly commit on success
+            success = true;
+            logger.debug("Operation completed successfully, connection committed");
+            return result;
+            
+        } catch (SQLException e) {
+            if (conn != null) {
+                try {
+                    conn.rollback();
+                    logger.debug("Transaction rolled back due to error");
+                } catch (SQLException rollbackEx) {
+                    logger.error("Failed to rollback transaction", rollbackEx);
+                }
+            }
+            throw e;
+            
+        } finally {
+            if (conn != null) {
+                try {
+                    conn.close();
+                    logger.debug("Connection closed");
+                } catch (SQLException e) {
+                    logger.error("Error closing connection", e);
+                }
+            }
+        }
+    }
+    
+    /**
+     * Executes a read-only operation with automatic connection management.
+     * This method uses read-only mode and auto-commit for better performance.
+     * 
+     * @param config the database configuration
+     * @param operation the operation to execute
+     * @param <T> the return type
+     * @return the result of the operation
+     * @throws SQLException if the operation fails
+     */
+    public static <T> T executeReadOnly(DatabaseConfig config, ConnectionOperation<T> operation) 
+            throws SQLException {
+        Connection conn = null;
+        
+        try {
+            conn = createConnection(config);
+            conn.setAutoCommit(true); // Use auto-commit for read-only
+            conn.setReadOnly(true);
+            
+            T result = operation.execute(conn);
+            logger.debug("Read-only operation completed successfully");
+            return result;
+            
+        } finally {
+            if (conn != null) {
+                try {
+                    conn.close();
+                    logger.debug("Read-only connection closed");
+                } catch (SQLException e) {
+                    logger.error("Error closing read-only connection", e);
+                }
+            }
+        }
+    }
+    
+    /**
+     * Tests if a database connection can be established with the given configuration.
+     * 
+     * @param jdbcUrl the JDBC URL to test
+     * @param username the database username
+     * @param password the database password
+     * @return true if connection successful, false otherwise
+     */
+    public static boolean testConnection(String jdbcUrl, String username, String password) {
+        Connection conn = null;
+        try {
+            conn = DriverManager.getConnection(jdbcUrl, username, password);
+            return conn.isValid(CONNECTION_TIMEOUT_SECONDS);
+        } catch (SQLException e) {
+            logger.debug("Connection test failed: {}", e.getMessage());
+            return false;
+        } finally {
+            if (conn != null) {
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    // Ignore close errors in test
+                }
+            }
+        }
+    }
+    
+    /**
+     * Checks if a Derby database is locked by another process.
+     * 
+     * @param databasePath the path to the database to check
+     * @return true if the database is locked, false if available
      */
     public static boolean isDerbyDatabaseLockedByAnotherProcess(String databasePath) {
-        if (databasePath == null || databasePath.trim().isEmpty()) {
-            throw new IllegalArgumentException("Database path cannot be null or empty");
-        }
-        
-        try (Connection testConnection = createQuickTestConnection(databasePath)) {
-            // Connection succeeded - database is available
-            return false;
+        Connection testConnection = null;
+        try {
+            String jdbcUrl = "jdbc:derby:" + databasePath;
+            testConnection = DriverManager.getConnection(jdbcUrl);
+            return false; // Successfully connected, not locked
         } catch (SQLException e) {
-            // Analyze the exception to determine if it's a lock-related issue
-            return isDerbyLockException(e);
+            // Check for Derby lock exception
+            String sqlState = e.getSQLState();
+            if ("XJ040".equals(sqlState) || "XSDB6".equals(sqlState)) {
+                logger.debug("Database is locked by another process: {}", databasePath);
+                return true;
+            }
+            // Other errors mean database doesn't exist or other issues
+            logger.debug("Database connection test failed for other reasons: {}", e.getMessage());
+            return false;
+        } finally {
+            if (testConnection != null) {
+                try {
+                    testConnection.close();
+                } catch (SQLException e) {
+                    // Ignore close errors
+                }
+            }
         }
     }
     
     /**
      * Tests if a database profile's configured database is currently available.
-     * 
-     * <p>This is a convenience method that extracts the database path from a
-     * DatabaseProfile and tests it for lock conditions. This method name clearly
-     * communicates its purpose in the context of profile management.
      * 
      * @param profile the database profile to test for availability
      * @return true if the profile's database is available, false if locked
@@ -93,81 +217,38 @@ public class DatabaseConnectionManager {
     }
     
     /**
-     * Creates a quick test connection to the specified Derby database.
+     * Provides user-friendly explanation for database connection failures.
      * 
-     * <p>This connection is designed for testing purposes only and includes:
-     * <ul>
-     *   <li>Short connection timeout to avoid blocking</li>
-     *   <li>No database creation if it doesn't exist</li>
-     *   <li>Minimal resource allocation for quick cleanup</li>
-     * </ul>
-     * 
-     * @param databasePath the path to the Derby database
-     * @return a test connection that should be closed immediately
-     * @throws SQLException if connection cannot be established
-     */
-    private static Connection createQuickTestConnection(String databasePath) throws SQLException {
-        String jdbcUrl = "jdbc:derby:" + databasePath + ";create=false";
-        
-        // Set a short timeout to avoid blocking the application startup
-        DriverManager.setLoginTimeout(CONNECTION_TIMEOUT_SECONDS);
-        
-        return DriverManager.getConnection(jdbcUrl, "", "");
-    }
-    
-    /**
-     * Recognizes Derby-specific error codes that indicate database lock conflicts.
-     * 
-     * <p>Derby uses specific SQL state codes and error messages to indicate different
-     * types of database lock conditions:
-     * <ul>
-     *   <li><strong>XJ040:</strong> Database cannot be accessed while in snapshot mode</li>
-     *   <li><strong>XJ041:</strong> Database already exists and is being used by another process</li>
-     *   <li><strong>XBM0J:</strong> Cannot create database because it's locked</li>
-     * </ul>
-     * 
-     * <p>This method also checks for common lock-related keywords in error messages
-     * to catch additional lock scenarios that might not use standard SQL states.
-     * 
-     * @param e the SQLException to analyze for lock indicators
-     * @return true if the exception indicates a database lock condition
-     */
-    private static boolean isDerbyLockException(SQLException e) {
-        String sqlState = e.getSQLState();
-        String message = e.getMessage().toLowerCase();
-        
-        // Check for specific Derby lock-related SQL states
-        if ("XJ040".equals(sqlState) || 
-            "XJ041".equals(sqlState) || 
-            "XBM0J".equals(sqlState)) {
-            return true;
-        }
-        
-        // Check for lock-related keywords in error messages
-        return message.contains("locked") ||
-               message.contains("another process") ||
-               message.contains("already in use") ||
-               message.contains("cannot access") ||
-               message.contains("database already exists");
-    }
-    
-    /**
-     * Provides a human-readable explanation of why a database connection failed.
-     * 
-     * <p>This method transforms technical Derby error messages into clear explanations
-     * that can be displayed to users or logged for troubleshooting. It helps bridge
-     * the gap between technical database errors and user-friendly communication.
-     * 
-     * @param e the SQLException that occurred during connection attempt
-     * @param databasePath the path that was being accessed when the error occurred
-     * @return a clear, human-readable explanation of the connection failure
+     * @param e the SQLException that occurred
+     * @param databasePath the database path that failed
+     * @return human-readable explanation of the error
      */
     public static String explainConnectionFailure(SQLException e, String databasePath) {
-        if (isDerbyLockException(e)) {
-            return "Database at '" + databasePath + "' is currently being used by another " +
-                   "MP3Org instance. The application will automatically use an alternative database.";
+        if (e == null) {
+            return "Unknown database error occurred";
         }
         
-        return "Cannot connect to database at '" + databasePath + "': " + e.getMessage();
+        String sqlState = e.getSQLState();
+        String message = e.getMessage();
+        
+        // Derby-specific error codes
+        if ("XJ040".equals(sqlState) || "XSDB6".equals(sqlState)) {
+            return "Database is locked by another instance of MP3Org. Please close other instances.";
+        } else if ("XJ004".equals(sqlState)) {
+            return "Database not found at: " + databasePath;
+        } else if ("08001".equals(sqlState)) {
+            return "Unable to connect to database. Check if the path exists: " + databasePath;
+        }
+        
+        // Generic message
+        return "Database error: " + message;
+    }
+    
+    /**
+     * Functional interface for database operations.
+     */
+    @FunctionalInterface
+    public interface ConnectionOperation<T> {
+        T execute(Connection conn) throws SQLException;
     }
 }