… src/main/java/org/hasting/util/DatabaseConnectionPool.java
… src/main/java/org/hasting/util/DatabaseManager.java